+++ mudock/include/mudock/compute/precomputed_adt_score.hpp
@@ -72,9 +72,9 @@ namespace mudock {
         std::memcpy(prot_min(), adt_prot.get_min_p(), 3 * sizeof(fp_type));
         std::memcpy(prot_max(), adt_prot.get_max_p(), 3 * sizeof(fp_type));
         std::memcpy(prot_center(), adt_prot.get_center_p(), 3 * sizeof(fp_type));
-        prot_index_x()[0]   = adt_prot.get_size_x();
-        prot_index_xy()[0]  = adt_prot.get_size_xy();
-        prot_index_xyz()[0] = adt_prot.get_size_xyz();
+        prot_index_x()[0]   = static_cast<int>(adt_prot.get_size_x());
+        prot_index_xy()[0]  = static_cast<int>(adt_prot.get_size_xy());
+        prot_index_xyz()[0] = static_cast<int>(adt_prot.get_size_xyz());
          std::memcpy(prot_grid_maps(),
                      adt_prot.get_maps_pointer(),
                      adt_prot.get_map_flat_size() * num_autodock_grids() * sizeof(fp_type));
@@ -165,7 +165,7 @@ namespace mudock {
         std::memcpy((void *) (nonbond_xB() + num_nonbond()[ligand_index]),
                     adt_ligand.non_bond_xB(),
                     non_bond_size * sizeof(int));
-        num_nonbond()[ligand_index + 1] = num_nonbond()[ligand_index] + non_bond_size;
+        num_nonbond()[ligand_index + 1] = static_cast<int>(num_nonbond()[ligand_index] + non_bond_size);
 
         //creazione mia nuova mappa fusa
         precomputed_protein my_fused_prot(grid_maps, sx, sy, sz, adt_ligand, ligand);
